[fix] Make jemalloc link properly with aml
Using both --with-jemalloc-prefix=jemk_ and --with-install-suffix=-aml
ensures that the libjemalloc we build for our internal use has unique
functions names and library names.

This patch fixes a problem with the libtool configuration, where we had
to link a static jemalloc library into the shared aml library to make
everything work. Instead, we can now use the proper LIBADD variables
with the unique library.

Note that with this patch the `make check` target requires a `make
install` beforehand.