[refactor/feature] change tilings to use uuids
Using variable arguments on the tile id for retrieving tiling info makes
the API difficult to use when more than one tile must be used at the
same time.

We change the API to use a tileid, with the assumption that any valuable
tiling will be able to define a workable uuid scheme.