[feature] add gitlab CI pipeline
Simple gitlab-ci config with a single step, running the full list of
configure, make, make install and make check.

No tags, no split build and test for now, as artifacts are a bit tricky
to get right.

This config should grow in the future to ensure that we run all the test
on all the platforms we want.