[ci] activate more warnings during compile
We've recently received messages indicating that some compiler
environments out there would choke on some of the code. Let's try to
make the code as C99 compliant and as reliable as possible by activating
more warnings and breaking the CI on it.

autoconf macros fail on -Werror -Wall because of unused variable
warnings, so we deactivate those.