[refactor] simplify CUDA detection and activation
The previous CUDA activation code was trying to differentiate between
building CUDA support and actually running the tests, but this isn't
really necessary: any platform with CUDA support should be able to run
the tests somewhere.

The configure.ac is modified to also check for all necessary cuda
runtime and headers at once.

We also introduce support for CUDA_HOME, which is an environment
variable available on some target systems.