[ci] single quote glob patterns when listing files
`git ls-files` accepts a list of patterns when listing files belonging to
the repository. If this pattern is interpreted as a glob by the shell,
it will however results in ls-files only listing the given filenames. To
avoid this issue, we need to systematically single quote the patterns
given to ls-files.

This is a difficult bug to uncover, as the shell will pass the pattern
as is if the pattern does match any files.

Note that the current gitlab-ci config does not have any issue, as there
are no .c or .h files in current working directory when these commands
are used.

Fix #60