[feature] add raw pointer access to layouts
In most application use cases, AML will need to pass a pointer to the
layout buffer for it to be handled by application kernel. That is to
say, we don't expect applications' inner code to use layout_deref in
tight loops. Thus, it should be possible to retrieve the underlying
buffer with the minimum amount of overhead.

Add an operator to retrieve this buffer, in the most straightforward way
possible.

See #61.

layout checkpatch