[feature] add raw pointer access to tilings
Same reasons as for aml_layout_rawptr.
Note that deref only works because we can be sure that none of our
tilings have padding or dummy data at the beginning of a tile. Otherwise
we would need to do a real index + rawptr on it, and clean up the layout
after.

tiling style