[feature] add function to print a layout
Add a generic API function to print the internal metadata of a layout,
not its contents.

The printing format is somewhat simple, one field per line, except for
arrays that are printed one row at a time. We print them in user-order.

The generic API follows the fprintf argument style, somewhat, while the
internal code follows the AML conventions.