[feature] add function to print area internals
Similar to layouts. I had to change the order/values of the CUDA flags
to ensure that I could fit all the "basic" flags in a small contiguous
array. Given the fact that MAPPED and GLOBAL are some kind of "extra"
option on the other flags, it might have made more sense to actual move
them into a different namespace/enum, but it will do in the meantime.