[fix] mmap tests fail on tricky system conditions
Some implementations require the file descriptor to be -1 when asking
for an anonymous mmap, making this test fail. The fix is to test both
anonymous mappings and file-backed mappings separately.

Some filesystems do not support the full posix range of file-backed
mmaps too (parallel file systems), so the file used for tests must be on
a simpler filesystem.