[fix] remove gettid from aml
gettid is a flaky linux-only symbol that might not be available on some
distros and is only used inside aml as part of a corner-case of
replicaset: to load balance between multiple replicas at the same
distance of a thread. Might as well make the load balance an arbitrary
decision and remove the symbol from the library.