[fix/ze] support a device with 0 subdevices
Fix allocation and typos.
Also add support for a ze device that successfully returns 0 subdevices
when asked.

Note that this allocation is not in line with the rest of the library,
and should be fixed at a later date. I've also let typos in the backend
header, should be fixed later.