[refactor] make ze dma thread-safe, simplify
This patch started as a way to make our ZE dma thread-safe, as the API
doesn't ensure safety when appending inside a command list.

A second reading of the docs lead us to a better implementation of our
synchronization primitives, for two reasons:
 - you can't create a command queue to try to synchronize the queue
   implicitly created by an immediate command list
 - you only need to send into the list a barrier event to wait for the
   entire list.

The resulting design use an event pool and a sync event per request that
the user is actually tracking, otherwise relies on barrier events. We
still have to lock a few things when actually sending requests.

This might not be completely efficient when creating massive amounts of
requests, but the assumption is that a client creating many requests at
once wouldn't track them all.