[feature] CI on nvhpc, the NVIDIA HPC compilers (#237)
Update excit to remove issues with hardcoded -Werror, test the code
against NVHPC SDK 24.7, and fix a couple of bugs in the process.

Note: src/features.c depends on CUDA when CUDA is active, so might as
well include the CUDA flags in the entire library.

Note: The hwloc tests are written to test hwloc or the platform, instead of
testing that aml is doing the right thing. We will need to change them
drastically.