[feature] Add MPI One-sided area
Add an area that can use one-sided window allocation to allocate memory.
The interface is limited by the required arguments on communicators,
requiring us to keep track of the created windows internally to be able
to cleanly destroy them.

Could probably also design a DMA to use puts and gets and fence, even if
that would be a bit strange in terms of API, since they don't really
match the "stream of copy request" model.